% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_arrays.R
\name{bind_array_helper}
\alias{bind_array_helper}
\title{Bind array helper}
\usage{
bind_array_helper(x, y)
}
\arguments{
\item{x}{Matrix or array 1}

\item{y}{Matrix or array 2}
}
\value{
Returns three-dimensional array containing elements x and y
}
\description{
Bind array helper
}
\keyword{internal}
