% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindArrays.R
\name{bindArrays}
\alias{bindArrays}
\title{Bind two arrays}
\usage{
bindArrays(array1, array2, message = TRUE)
}
\arguments{
\item{array1}{Three-dimensional array}

\item{array2}{Three-dimensional array}

\item{message}{Give warning if arrays are named inconsistently}
}
\value{
Returns new array
}
\description{
Bind two arrays
}
