% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avg.R
\name{calc_avg}
\alias{calc_avg}
\title{Calculate averageness}
\usage{
calc_avg(data)
}
\arguments{
\item{data}{Facefuns object or three-dimensional array of dimensions p, k (2 or 3), and n (minimum = 2)}
}
\value{
Returns tibble with distinctiveness and averageness (reversed distinctiveness) scores
}
\description{
Functions calculates averageness/distinctiveness as each template's distance from sample average
}
\examples{
data(LondonSet_aligned)
data(mirroredlandmarks)
calc_avg(LondonSet_aligned)
}
