% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fa.R
\name{calc_fa}
\alias{calc_fa}
\title{Calculate fluctuating asymmetry}
\usage{
calc_fa(data, mirroredlandmarks)
}
\arguments{
\item{data}{\code{facefuns} object or three-dimensional array of dimensions p, k, and n}

\item{mirroredlandmarks}{Vector specifying order of mirrored landmarks}
}
\value{
Returns tibble containing ID and FA scores
}
\description{
\lifecycle{maturing}

Calculates fluctuating asymmetry (FA) of landmark templates. FA scores are calculated as Euclidean distance between original and symmetrised templates, corrected for directional asymmetry of sample
}
\examples{
data(LondonSet_aligned)
data(mirroredlandmarks)
calc_fa(LondonSet_aligned, mirroredlandmarks)

}
