% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fageo.R
\name{calc_fageo}
\alias{calc_fageo}
\title{Calculate fluctuating asymmetry}
\usage{
calc_fageo(data, mirroredlandmarks)
}
\arguments{
\item{data}{\code{facefuns} object or three-dimensional array of dimensions p, k, and n}

\item{mirroredlandmarks}{Vector specifying index of mirrored landmarks}
}
\value{
Returns tibble containing ID and FA values
}
\description{
\lifecycle{experimental}

Function uses \link[geomorph]{bilat.symmetry} to calculate fluctuating asymmetry (FA) component for each template. FA score is calculated as distance between FA component and mean symmetric shape
}
\examples{
data(LondonSet_aligned)
data(mirroredlandmarks)
calc_fageo(LondonSet_aligned, mirroredlandmarks)

}
