% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vs.R
\name{calc_vs}
\alias{calc_vs}
\title{Calculate vector score}
\usage{
calc_vs(data, anchor1, anchor2)
}
\arguments{
\item{data}{PC scores of face(s) for which vector scores are to be calculated}

\item{anchor1}{PC scores of face(s) which will constitute lower anchor point}

\item{anchor2}{PC scores of face(s) which will constitute upper anchor point}
}
\value{
Returns tibble with columns "id" and "VS". If data contained rownames, these will be saved as ids.
}
\description{
Function projects data onto vector from anchor one (score of 0) to anchor two (score of 1). For all three arguments, input must be two-dimensional data frame or matrix with rows = specimens, and columns = PC scores.
}
\examples{
# CALCULATE FEMALE-MALE VECTOR SCORES
data("LondonSet_scores")
data("LondonSet_info")

fem <- LondonSet_scores \%>\%
  dplyr::filter(row.names(LondonSet_scores) \%in\%
  LondonSet_info$face_id[which(LondonSet_info$face_sex == "female")])

mal <- LondonSet_scores \%>\%
  dplyr::filter(row.names(LondonSet_scores) \%in\%
  LondonSet_info$face_id[which(LondonSet_info$face_sex == "male")])

calc_vs(LondonSet_scores, fem, mal)

}
