% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAS.R
\name{calcAS}
\alias{calcAS}
\title{Calculate asymmetry}
\usage{
calcAS(data, mirr_lms)
}
\arguments{
\item{data}{Quickstart object or three-dimensional array of dimensions p, k, and n}

\item{mirr_lms}{Vector specifying order of mirrored landmarks}
}
\value{
Returns tibble containing ID and asymmetry values
}
\description{
Function mirrors templates and runs GPA on original plus mirrored templates. Asymmetry scores are then calculated as Euclidean distance between original template and its symmetrized version (i.e. the mean of original and mirrored template)
}
\details{
NOTE: does not distinguish between directional and fluctuating asymmetry
}
\examples{
data(LondonSet_aligned)
data(mirr_lms)
calcAS(LondonSet_aligned, mirr_lms)

}
