% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAVG.R
\name{calcAVG}
\alias{calcAVG}
\title{Calculate averageness}
\usage{
calcAVG(data)
}
\arguments{
\item{data}{Facefuns object or three-dimensional array of dimensions p, k (2 or 3), and n (minimum = 2)}
}
\value{
Returns tibble with distinctiveness and averageness (reversed distinctiveness) scores
}
\description{
Functions calculates averageness/distinctiveness as each template's distance from sample average
}
