% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDS.R
\name{calcDS}
\alias{calcDS}
\title{Calculate linear discriminant scores}
\usage{
calcDS(data, group_info = 1, acc = FALSE)
}
\arguments{
\item{data}{Matrix or data frame containing PC scores of specimen (rows = specimen, cols = PCs)}

\item{group_info}{Numeric index of column in data that indicates actual group membership}

\item{acc}{Print accuracy of group membership prediction}
}
\value{
Returns tibble with columns "id" and "discrimScore". If data contained rownames, these will be saved as ids. If acc = TRUE, output will be list, with tibble "ds" containing id and vector score, and list "acc" containing accuracy info.
}
\description{
Function uses stepwise LDA to choose best predictors for group membership, and then "regular" LDA to predict discriminant scores.
}
\details{
First uses \link[klaR]{greedy.wilks} to perform stepwise forward variable selection for classification, then uses selected variables in \link[MASS]{lda}
}
