% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFA.R
\name{calcFA}
\alias{calcFA}
\title{Calculate fluctuating asymmetry}
\usage{
calcFA(data, mirr_lms)
}
\arguments{
\item{data}{Quickstart object or three-dimensional array of dimensions p, k, and n}

\item{mirr_lms}{Vector specifying order of mirrored landmarks}
}
\value{
Returns tibble containing ID and FA scores
}
\description{
Calculates fluctuating asymmetry (FA) of landmark templates. FA scores are calculated as Euclidean distance between original and symmetrized templates, corrected for sample directional asymmetry
}
\examples{
data(LondonSet_aligned)
data(mirr_lms)
calcFA(LondonSet_aligned, mirr_lms)

}
