% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFA_geo.R
\name{calcFA_geo}
\alias{calcFA_geo}
\title{Calculate fluctuating asymmetry}
\usage{
calcFA_geo(data, mirr_lms)
}
\arguments{
\item{data}{Quickstart object or three-dimensional array of dimensions p, k, and n}

\item{mirr_lms}{Vector specifying order of mirrored landmarks}
}
\value{
Returns tibble containing ID and asymmetry values
}
\description{
Function uses \link[geomorph]{bilat.symmetry} to calculate fluctuating asymmetry component for each template. FA score is calculated as distance between FA component and mean symmetric shape
}
\examples{
data(LondonSet_aligned)
data(mirr_lms)
calcFA(LondonSet_aligned, mirr_lms)

}
