% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShapeVS.R
\name{calcShapeVS}
\alias{calcShapeVS}
\title{Calculate vector scores from Procrustes-aligned landmark templates}
\usage{
calcShapeVS(data, anchor1_index, anchor2_index, symm = FALSE, mirr_lms)
}
\arguments{
\item{data}{Output from \link[facefuns]{quickstart}}

\item{anchor1_index}{Vector specifying indices of faces which will constitute lower anchor point}

\item{anchor2_index}{Vector specifying indices of faces which will constitute upper anchor point}

\item{symm}{Symmetrize templates prior to calculating vector scores}

\item{mirr_lms}{Vector specifiying order of mirrored landmarks (required for symmetrizing)}
}
\value{
Returns tibble with columns "id" and "VS". If data contained rownames, these will be saved as ids.
}
\description{
Function projects data onto vector from anchor one (score of 0) to anchor two (score of 1). See \link[facefuns]{calcVS}
}
\examples{
# CALCULATE FEMALE-MALE VECTOR SCORES FROM (SYMMETRIZED) SET OF FACES
# Read and prep data
path_to_tem <- system.file("extdata", "tem", package="facefuns")
remove_points <- c(45:50, 100:104, 116:125, 146:158, 159:164, 165:170, 171:174, 175:179, 184:185)

data <- read_lmdata(lmdata = path_to_tem,
                    remove_points = remove_points)

shapedata <- quickstart(data = data,
                       pc_criterion = "broken_stick",
                       plot_sample = FALSE,
                       message = FALSE)

# Load info on sex of faces
data("LondonSet_info")

# Load vector specifying order of mirrored landmarks
data("mirr_lms")

# Pull together indices of faces for male and female averages
fem_i <- gsub("^ID=","", dimnames(shapedata$array)[[3]]) \%in\%
LondonSet_info$face_id[which(LondonSet_info$face_sex == "female")]

mal_i <- gsub("^ID=", "", dimnames(shapedata$array)[[3]]) \%in\%
LondonSet_info$face_id[which(LondonSet_info$face_sex == "male")]

# Calculate vector scores
calcShapeVS(shapedata, fem_i, mal_i, symm=TRUE, mirr_lms = mirr_lms)

}
