% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVS.R
\name{calcVS}
\alias{calcVS}
\title{Calculate vector score}
\usage{
calcVS(data, anchor1, anchor2)
}
\arguments{
\item{data}{PC scores of face(s) for which vector scores are to be calculated}

\item{anchor1}{PC scores of face(s) which will constitute lower anchor point}

\item{anchor2}{PC scores of face(s) which will constitute upper anchor point}
}
\value{
Returns tibble with columns "id" and "vectorScore". If data contained rownames, these will be saved as ids.
}
\description{
Function projects data onto vector from anchor one (score of 0) to anchor two (score of 1). For all three arguments, input must be two-dimensional matrix with rows = specimen, and columns = PC scores.
}
