% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_array_to_matrix.R
\name{convert_array_to_matrix}
\alias{convert_array_to_matrix}
\title{Convert 3D array into matrix}
\usage{
convert_array_to_matrix(x)
}
\arguments{
\item{x}{A three-dimensional array (p x k x n) of landmark coordinates}
}
\value{
Returns two-dimensional matrix with n rows and p x k columns. If input has labels, these will be preserved.
}
\description{
Convert 3D array into matrix
}
\examples{
data("LondonSet_aligned")
data_matrix <- convert_array_to_matrix(LondonSet_aligned)
}
