% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_points_to_mesh.R
\name{convert_points_to_mesh}
\alias{convert_points_to_mesh}
\title{Convert 3-D coordinates to 3-D mesh}
\usage{
convert_points_to_mesh(point_matrix, alpha, showme = TRUE)
}
\arguments{
\item{point_matrix}{Coordinate data as matrix (individual points, x/y/z coordinates)}

\item{alpha}{See \link[alphashape3d]{ashape3d}}

\item{showme}{Return default 3D plot}
}
\value{
3D mesh
}
\description{
Computes alpha-shape of three-dimensional coordinates, see \href{https://github.com/cran/alphashape3d}{here} for more info
}
\examples{

\dontrun{

# CREATE EXAMPLE POINT MATRIX
path_to_ply <- system.file("extdata", "ply", package="facefuns")
data <- read_vertices(path_to_ply)

shapedata <- facefuns(data = data,
                      pc_criterion = "broken_stick",
                      quiet = TRUE)

point_matrix <- shapedata$pc_plot$PC1$minus

# RUN FUNCTION
mesh <- convert_points_to_mesh(point_matrix = point_matrix,
                               alpha = 1.2,
                               showme = FALSE)

# PLOT RESULT
rgl::open3d()
rgl::plot3d(mesh,
            type = "wire", col = "gray",
            box = FALSE, axes = FALSE,
            xlab = "", ylab = "", zlab = "",
            specular = "black")
rgl::aspect3d("iso")
}


}
