% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_stimlist_to_array.R
\name{convert_stimlist_to_array}
\alias{convert_stimlist_to_array}
\title{Convert webmorph stimlist to array (borrowed from webmorph::tems_to_array)}
\usage{
convert_stimlist_to_array(stimlist)
}
\arguments{
\item{stimlist}{Object of class webmorph_list}
}
\value{
Returns 3D array containing template files
}
\description{
Convert webmorph stimlist to array (borrowed from webmorph::tems_to_array)
}
\author{
Lisa DeBruine
}
\keyword{internal}
