% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertArrayToMatrix.R
\name{convertArrayToMatrix}
\alias{convertArrayToMatrix}
\title{Convert 3D array into matrix}
\usage{
convertArrayToMatrix(x)
}
\arguments{
\item{x}{A three-dimensional array (p x k x n) of landmark coordinates}
}
\value{
Returns two-dimensional matrix with n rows and p x k columns. If input has labels, these will be preserved.
}
\description{
Convert 3D array into matrix
}
\examples{
data("LondonSet_aligned")
data_matrix <- convertArrayToMatrix(LondonSet_aligned)
}
