% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPointsToMesh.R
\name{convertPointsToMesh}
\alias{convertPointsToMesh}
\title{Convert 3D coordinates to 3D mesh}
\usage{
convertPointsToMesh(point.matrix, alpha, showme = TRUE)
}
\arguments{
\item{point.matrix}{Coordinate data as matrix (individual points, x/y/z coordinates)}

\item{alpha}{See \link[alphashape3d]{ashape3d}}

\item{showme}{Return default 3D plot}
}
\value{
3D mesh
}
\description{
Computes alpha-shape of three-dimensional coordinates, see \href{https://github.com/cran/alphashape3d}{here} for more info.
}
\examples{

data(example_3D_face)

M.mesh <- convertPointsToMesh(point.matrix = example_3D_face,
alpha = 1.2,
showme = FALSE)

# Show 3D face in R notebook
rgl::open3d()
rgl::plot3d(M.mesh,
     type = "wire", col = "gray",
      box = FALSE, axes = FALSE,
      xlab = "", ylab = "", zlab = "",
      specular = "black")
rgl::aspect3d("iso")
rgl::rglwidget()

}
