% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStimlistToArray.R
\name{convertStimlistToArray}
\alias{convertStimlistToArray}
\title{Convert webmorph stimlist to array (borrowed from webmorph::tems_to_array)}
\usage{
convertStimlistToArray(stimlist)
}
\arguments{
\item{stimlist}{Object of class webmorph_list}
}
\value{
Returns 3D array containing template files
}
\description{
Convert webmorph stimlist to array (borrowed from webmorph::tems_to_array)
}
\author{
Lisa DeBruine
}
