% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_3D_face.R
\docType{data}
\name{example_3D_face}
\alias{example_3D_face}
\title{3D example}
\format{
fill me in
}
\source{
\href{http://facelab.org/}{Figshare}
}
\usage{
data(example_3D_face)
}
\description{
Shape coordinates of a 3D composite face
}
\examples{
data(example_3D_face)
}
\references{
fill me in (\href{http://facelab.org/}{Facelab})
}
\keyword{datasets}
