% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facefuns_app.R
\name{facefuns_app}
\alias{facefuns_app}
\title{Launch Shiny App}
\usage{
facefuns_app(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch a  shiny app that runs locally in RStudio or your web browser (recommended). It does not connect to the web at all, so your data are completely private.
}
\examples{
\dontrun{ facefuns_app() }

}
