% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frl_features.R
\name{frl_features}
\alias{frl_features}
\title{Get point indices for FRL standard template features}
\usage{
frl_features(...)
}
\arguments{
\item{...}{Vector of feature names}
}
\value{
Vector of corresponding FRL template indices
}
\description{
\lifecycle{experimental}

Available features in alphabetical order: cheekbones, chin, ears, face, halo, left_brow, left_eye, lowerlip, mouth, neck1, neck2, nose, philtrum, right_brow, right_eye, smile_lines, undereyes. You can also use "frlshape2019" to choose features used in \href{https://doi.org/10.1037/xhp0000685}{Holzleitner et al., 2019}
}
\examples{
frl_features("mouth")
frl_features("left_eye", "right_eye")
frl_features("frlshape2019")

}
\author{
Lisa DeBruine
}
