% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{Flat violin geom by Ben Marwick}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{A value}

\item{data}{A value}

\item{stat}{A value}

\item{position}{A value}

\item{trim}{A value}

\item{scale}{A value}

\item{show.legend}{A value}

\item{inherit.aes}{A value}

\item{...}{A value}
}
\description{
Create flat violin plot with ggplot2 with geom_flat_violin. Shamelessly copy-pasted from \href{https://gist.githubusercontent.com/benmarwick/2a1bb0133ff568cbe28d}{Ben Marwick}.
}
\examples{

ggplot2::ggplot(ggplot2::diamonds, ggplot2::aes(cut, carat)) +
 ggplot2::coord_flip() +
 geom_flat_violin()

}
