% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_shape_array.R
\name{is_shape_array}
\alias{is_shape_array}
\title{Helper function to check if input is array of right dimensions for shape analyses}
\usage{
is_shape_array(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE or FALSE
}
\description{
Helper function to check if input is array of right dimensions for shape analyses
}
\keyword{internal}
