% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LondonSet_info.R
\docType{data}
\name{LondonSet_info}
\alias{LondonSet_info}
\title{FRL London Set}
\format{
fill me in
}
\source{
\href{https://figshare.com/articles/Face_Research_Lab_London_Set/5047666b}{Figshare}
}
\usage{
data(LondonSet_info)
}
\description{
Face Research Lab London face image data set. Images are of 102 adult faces that were originally delineated with 189 landmarks. Original template files were converted to .tps file format and some landmark points removed for shape analyses (see https://osf.io/98qf4/)
}
\details{
This file contains info on age, gender and ethnicity of faces
}
\examples{
data(LondonSet_info)
}
\references{
fill me in
(\href{https://figshare.com/articles/Face_Research_Lab_London_Set/5047666}{Figshare})
}
\keyword{datasets}
