% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make2DPCs.R
\name{make2DPCs}
\alias{make2DPCs}
\title{Create list of 2D PCs}
\usage{
make2DPCs(pca_output, ref, which_pcs, vis_sd)
}
\arguments{
\item{pca_output}{Output from \code{gm.prcomp}}

\item{ref}{Reference face}

\item{which_pcs}{Which PCs are to be created. Single number or vector, maximum allowed is 3}

\item{vis_sd}{Extent of desired manipulation in units of standard deviation}
}
\value{
A list of 2D coordinates for the reference face at \code{-vis_SD} and |code{vis_SD} standard deviations for each principal component
}
\description{
Create list of 2D PCs
}
\examples{
data(LondonSet_aligned)
data_aligned <- LondonSet_aligned
pca_output <- geomorph::gm.prcomp(data_aligned)
ref <- geomorph::mshape(data_aligned)
make2DPCs(pca_output, ref, which_pcs = 1:3, vis_sd = 1.5)

}
