% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_3dpcs.R
\name{make_3dpcs}
\alias{make_3dpcs}
\title{Make and plot 3D principal components}
\usage{
make_3dpcs(
  pca_output,
  numberOfPCs,
  visSD,
  mean_shape,
  plot = TRUE,
  return = TRUE
)
}
\arguments{
\item{pca_output}{Prcomp output}

\item{numberOfPCs}{How many PCs to visualize}

\item{visSD}{How many standard deviations to visualize, e.g. 3}

\item{mean_shape}{Reference coordinates on which visualization will be based}

\item{plot}{Plot or not}

\item{return}{Return list of PCs}
}
\value{
List of 3D meshes visualizing PCs according to requested SDs; e.g., if numberOfPCs was set to 3, six meshes will be returned - PCs 1, 2, and 3 at plus and minus visSD
}
\description{
\lifecycle{experimental}
}
