% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirr_lms.R
\docType{data}
\name{mirr_lms}
\alias{mirr_lms}
\title{Mirrored landmarks order}
\format{
fill me in
}
\usage{
data(mirr_lms)
}
\description{
Textfile re-specifying landmark order for mirroring reduced FRL default template of 132 points.
}
\examples{
data(mirr_lms)
}
\keyword{datasets}
