% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_templates.R
\name{mirror_templates}
\alias{mirror_templates}
\title{Mirror landmark templates}
\usage{
mirror_templates(data, mirroredlandmarks)
}
\arguments{
\item{data}{Array holding landmark data of dimensions p, k, and n. p = number of landmarks, k = dimension (2D or 3D), n = number of specimens.}

\item{mirroredlandmarks}{Vector specifying order of mirrored landmarks}
}
\value{
Returns array containing mirrored landmark data
}
\description{
Mirror landmark templates
}
\examples{
data(LondonSet_aligned)
data(mirroredlandmarks)

mirr <- mirror_templates(LondonSet_aligned, mirroredlandmarks)

geomorph::plotAllSpecimens(LondonSet_aligned)
geomorph::plotAllSpecimens(mirr)

}
