% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirroredlandmarks.R
\docType{data}
\name{mirroredlandmarks}
\alias{mirroredlandmarks}
\title{Mirrored landmarks order}
\format{
fill me in
}
\usage{
data(mirroredlandmarks)
}
\description{
Textfile re-specifying landmark order for mirroring reduced FRL default template of 132 points.
}
\examples{
data(mirroredlandmarks)
}
\keyword{datasets}
