% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorTemplates.R
\name{mirrorTemplates}
\alias{mirrorTemplates}
\title{Mirror landmark templates}
\usage{
mirrorTemplates(data, mirr_lms)
}
\arguments{
\item{data}{Array holding landmark data of dimensions p, k, and n. p = number of landmarks, k = dimension (2D or 3D), n = number of specimens.}

\item{mirr_lms}{Vector specifying order of mirrored landmarks}
}
\value{
Returns array containing mirrored landmark data
}
\description{
Mirror landmark templates
}
\examples{
data(LondonSet_aligned)
data(mirr_lms)

mirr <- mirrorTemplates(LondonSet_aligned, mirr_lms)

geomorph::plotAllSpecimens(LondonSet_aligned)
geomorph::plotAllSpecimens(mirr)

}
