% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2dpcs.R
\name{pc_pair_plot}
\alias{pc_pair_plot}
\title{pc_pair_plot}
\usage{
pc_pair_plot(f1, f2, pc)
}
\arguments{
\item{f1}{Function for first plot}

\item{f2}{Function for second plot}

\item{pc}{Numeric label}
}
\value{
plot
}
\description{
Set up specs for plotting pairs of PCs
}
