% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2DPCs.R
\name{plot2DPCs}
\alias{plot2DPCs}
\title{Plot 2D PCs}
\usage{
plot2DPCs(input, ref, which_pcs = 1:3, vis_sd = 3)
}
\arguments{
\item{input}{Output from \code{gm.prcomp} or list of PC coordinates}

\item{ref}{Reference face (sample average is recommended)}

\item{which_pcs}{Which PCs are to be created. Single number or vector, maximum allowed is 5}

\item{vis_sd}{Extent of desired manipulation in units of standard deviation}
}
\value{
Returns plot
}
\description{
Convenience function for plotting shape PCs based on \code{geomorph::plotRefToTarget()}. Other than \code{plotRefToTarget}, magnitude of visualized difference in shape between reference face at average and low/high levels of the respective PC is based on standard deviations (instead of range).
}
\examples{
data(LondonSet_aligned)
pca_output <- geomorph::gm.prcomp(LondonSet_aligned)
ref <- geomorph::mshape(LondonSet_aligned)
plot2DPCs(pca_output, ref)

}
