% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2dpcs.R
\name{plot_2dpcs}
\alias{plot_2dpcs}
\title{Plot 2D PCs}
\usage{
plot_2dpcs(input, ref = NULL, which_pcs = 1:3, vis_sd = 3, output = NULL)
}
\arguments{
\item{input}{Object of class \code{facefuns}, \code{prcomp}, \code{gm.prcomp} or PC_list (see \code{make_pcs})}

\item{ref}{Reference face. Defaults to sample average if input is facefuns object}

\item{which_pcs}{Which PCs are to be created. Single number or vector, maximum allowed is 5}

\item{vis_sd}{Extent of desired manipulation in units of standard deviation}

\item{output}{Specify file name/path to save plot}
}
\value{
Returns plot
}
\description{
Convenience function for plotting shape PCs based on \code{geomorph::plotRefToTarget()}. Other than \code{plotRefToTarget}, magnitude of visualized difference in shape between reference face at average and low/high levels of the respective PC is based on standard deviations (instead of range).
}
\examples{
data(LondonSet_aligned)
pca_output <- geomorph::gm.prcomp(LondonSet_aligned)
ref <- geomorph::mshape(LondonSet_aligned)
plot_2dpcs(pca_output, ref)

}
