% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3dpcs.R
\name{plot_3dpcs}
\alias{plot_3dpcs}
\title{Plot 3D PCs}
\usage{
plot_3dpcs(input, ref, which_pcs = 1, vis_sd = 3, alpha = 1.2)
}
\arguments{
\item{input}{Object of class \code{prcomp}, \code{gm.prcomp} or PC_list (see \code{make_pcs})}

\item{ref}{Reference face (sample average is recommended)}

\item{which_pcs}{Which PCs are to be created. Single number or vector, maximum allowed is 3}

\item{vis_sd}{Extent of desired manipulation in units of standard deviation}

\item{alpha}{See \link[alphashape3d]{ashape3d}}
}
\value{
Returns plot
}
\description{
\lifecycle{experimental}
}
\examples{

\dontrun{
path_to_ply <- system.file("extdata", "ply", package="facefuns")
data <- read_vertices(path_to_ply)
shapedata <- facefuns3d(data = data,
pc_criterion = "broken_stick")

plot_3dpcs(input = shapedata$pc_plot, which_pcs = 1, ref = shapedata$average, alpha = 2)

}

}
