% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDS.R
\name{print.ds_out}
\alias{print.ds_out}
\title{Print for ds_out}
\usage{
\method{print}{ds_out}(x, ...)
}
\arguments{
\item{x}{a list of class ds_out}

\item{...}{arguments passed to or from other methods}
}
\value{
prints ds_out
}
\description{
Print for ds_out
}
