% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facefuns2d.R, R/facefuns3d.R
\name{print.facefuns_obj}
\alias{print.facefuns_obj}
\title{Print for facefuns_obj}
\usage{
\method{print}{facefuns_obj}(x, ...)

\method{print}{facefuns_obj}(x, ...)
}
\arguments{
\item{x}{a list of class facefuns_obj}

\item{...}{arguments passed to or from other methods}
}
\value{
prints summary

prints summary
}
\description{
Print for facefuns_obj

Print for facefuns_obj
}
\keyword{internal}
