% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstart.R
\name{print.quickstart}
\alias{print.quickstart}
\title{Print for quickstart}
\usage{
\method{print}{quickstart}(x, ...)
}
\arguments{
\item{x}{a list of class quickstart}

\item{...}{arguments passed to or from other methods}
}
\value{
prints summary
}
\description{
Print for quickstart
}
