% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R, R/selectPCs.R
\name{print.selected_PCs}
\alias{print.selected_PCs}
\title{Print for basic_output}
\usage{
\method{print}{selected_PCs}(x, ...)

\method{print}{selected_PCs}(x, ...)
}
\arguments{
\item{x}{a list of class selected_PCs}

\item{...}{arguments passed to or from other methods}
}
\value{
prints the basic output

prints the selected PCs
}
\description{
Print for basic_output

Print for selected_PCs
}
