% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ply.R
\name{read_ply}
\alias{read_ply}
\title{Read vertices from binary PLY files}
\usage{
read_ply(plydata)
}
\arguments{
\item{plydata}{Path to single PLY or directory containing PLY files}
}
\value{
Returns three-dimensional array of dimensions p, k, and n. p = number of landmarks, k = dimension (3D, as opposed to 2D), n = number of specimens
}
\description{
\lifecycle{experimental}

Uses \link[Morpho]{file2mesh} to read binary_little_endian PLY files and saves vertex data into three-dimensional array
}
\examples{

path_to_ply <- system.file("extdata", "ply", package="facefuns")
data <- read_ply(plydata = path_to_ply)

}
