% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vertices.R
\name{read_vertices}
\alias{read_vertices}
\title{Read vertices from PLY or OBJ files}
\usage{
read_vertices(meshdata)
}
\arguments{
\item{meshdata}{Path to single PLY/OBJ or directory containing PLY/OBJ files. If both files types are present, PLYs will be prioritised}
}
\value{
Returns array of dimensions p, k, and n. p = number of landmarks, k = dimension of data, n = number of specimens
}
\description{
\lifecycle{experimental}

Uses \link[Morpho]{file2mesh} to read PLY files or \link{readOBJ} to read OBJ files, and saves vertex data into three-dimensional array
}
\examples{

path_to_ply <- system.file("extdata", "ply", package="facefuns")
data <- read_vertices(meshdata = path_to_ply)

path_to_obj <- system.file("extdata", "obj", package="facefuns")
data <- read_vertices(meshdata = path_to_obj)

}
