% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_points.R
\name{remove_points}
\alias{remove_points}
\title{Remove points/landmarks from 2-D templates}
\usage{
remove_points(
  lmdata,
  points,
  relabel_points = TRUE,
  plot = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{lmdata}{Array of dimensions p (number of landmarks) x k (dimensionality of data, only 2 allowed) x n (number of specimens)}

\item{points}{You can specify points as numeric vector. If your original templates are standard FRL templates (p=189), you can also specify features by their FRL names (see \link[facefuns]{frl_features})}

\item{relabel_points}{If false, old landmark labels will be retained}

\item{plot}{Plot new landmark templates}

\item{quiet}{Print how many points were deleted}
}
\value{
Array of new dimensions p (reduced number of landmarks) x k x n
}
\description{
\lifecycle{experimental}

Delete points/landmarks from array of 2-D landmark templates
}
\examples{

path_to_tem <- system.file("extdata", "tem", package="facefuns")
data <- read_lmdata(lmdata = path_to_tem,
                    plot = TRUE)

new_data <- remove_points(lmdata = data,
                          points = "frlshape2019",
                          plot = TRUE)

}
