% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pcs.R
\name{select_pcs}
\alias{select_pcs}
\title{Select principal components}
\usage{
select_pcs(
  pca_output,
  method = c("broken_stick", "kaiser_guttman", "total_variance"),
  total_var = 0.95
)
}
\arguments{
\item{pca_output}{Output from \code{gm.prcomp} or \code{prcomp}}

\item{method}{Method used to select PCs. Defaults to the most conservative "broken_stick". Other methods are "kaiser_guttman" (PCs with eigenvalues greater than the mean eigenvalue) and "total_variance" (PCs explaining at least \code{total_var} variance)}

\item{total_var}{Total variance to choose for "total_variance" method. Defaults to .95}
}
\value{
Returns selected PCs and variance explained
}
\description{
\code{select_pcs} returns a list of principal components as chosen by one of three methods (from D.A. Jackson (1993). Ecology, 74, 2204-2214).
}
\note{
Code for "broken_stick" method adapted from function \href{http://adn.biol.umontreal.ca/~numericalecology/numecolR/}{"evplot" by Francois Gillet}.
}
