% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symm_templates.R
\name{symm_templates}
\alias{symm_templates}
\title{Symmetrize landmark templates}
\usage{
symm_templates(data, mirroredlandmarks)
}
\arguments{
\item{data}{Array holding landmark data of dimensions p, k, and n. p = number of landmarks, k = dimension (2D or 3D), n = number of specimens.}

\item{mirroredlandmarks}{Vector specifying order of mirrored landmarks}
}
\value{
Returns array containing symmetrized landmark data
}
\description{
Mirrors landmark templates for each specimen and calculates mean of original and mirrored template (see also \link[facefuns]{mirror_templates})
}
\examples{

data(LondonSet_aligned)
data(mirroredlandmarks)

symm <- symm_templates(LondonSet_aligned, mirroredlandmarks)

geomorph::plotAllSpecimens(LondonSet_aligned)
geomorph::plotAllSpecimens(symm)

}
