% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmTemplates.R
\name{symmTemplates}
\alias{symmTemplates}
\title{Symmetrize landmark templates}
\usage{
symmTemplates(data, mirr_lms)
}
\arguments{
\item{data}{Array holding landmark data of dimensions p, k, and n. p = number of landmarks, k = dimension (2D or 3D), n = number of specimens.}

\item{mirr_lms}{Vector specifying order of mirrored landmarks}
}
\value{
Returns array containing symmetrized landmark data
}
\description{
Mirrors landmark templates for each specimen and calculates mean of original and mirrored template (see also \link[facefuns]{mirrorTemplates})
}
\examples{
data(LondonSet_aligned)
data(mirr_lms)

symm <- symmTemplates(LondonSet_aligned, mirr_lms)

geomorph::plotAllSpecimens(LondonSet_aligned)
geomorph::plotAllSpecimens(symm)
}
