% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stimlist_to_tps.R
\name{write_stimlist_to_tps}
\alias{write_stimlist_to_tps}
\title{Write webmorph stimlist to TPS (borrowed from webmorph::write_tps)}
\usage{
write_stimlist_to_tps(stimlist, path_to_tps = NULL)
}
\arguments{
\item{stimlist}{Object of class webmorph_list}

\item{path_to_tps}{Optional file name of TPS file}
}
\value{
Returns webmorph_list in TPS format
}
\description{
Write webmorph stimlist to TPS (borrowed from webmorph::write_tps)
}
\author{
Lisa DeBruine
}
\keyword{internal}
