% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStimlistToTPS.R
\name{writeStimlistToTPS}
\alias{writeStimlistToTPS}
\title{Write webmorph stimlist to TPS (borrowed from webmorph::write_tps)}
\usage{
writeStimlistToTPS(stimlist, path_to_tps = NULL)
}
\arguments{
\item{stimlist}{Object of class webmorph_list}

\item{path_to_tps}{Optional file name of TPS file}
}
\value{
Returns webmorph_list in TPS format
}
\description{
Write webmorph stimlist to TPS (borrowed from webmorph::write_tps)
}
\author{
Lisa DeBruine
}
