% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{basics}
\alias{basics}
\title{Basic output for face shape data}
\usage{
basics(
  path_to_data,
  p = 189,
  remove_points = NA,
  rotate = c(NA, "flipX", "flipY", "rotateC", "rotateCC"),
  plot = TRUE,
  pc_criterion = "broken_stick"
)
}
\arguments{
\item{path_to_data}{Path to .tps file or directory containing .tem files}

\item{p}{Number of landmarks images were delineated with. Defaults to value for FRL standard template.}

\item{remove_points}{If reading from template files, it can be specified whether any landmarks should be removed. If no value is entered, all landmarks are retained.}

\item{rotate}{The type of rotation or flip to be performed. Specimens can be flipped with respect to x or y axes, or rotated clockwise or counter-clockwise. See \code{geomorphs}'s \link[geomorph]{rotate.coords}}

\item{plot}{Plot landmark coordinates of sample. See \code{geomorphs}'s \link[geomorph]{plotAllSpecimens}}

\item{pc_criterion}{Criterion used to choose which PCs to retain. See \link[facefuns]{selectPCs}}
}
\value{
Returns a list of the following:
\item{data}{Three-dimensional array containing original data}
\item{aligned}{Three-dimensional array containing Procrustes-aligned/rotated data}
\item{scores}{Principal component scores for each face}
\item{pcs}{List of selected PCs, their importance and method used to select PCs}
\item{ref}{Coordinates of sample average}
\item{plot}{Coordinates of sample average at -3SD and +3SD for each selected PC}
}
\description{
Function performs some of the routine steps for getting landmark data ready for shape analyses. Data can be imported from a directory of Webmorph template files, or by reading an existing TPS file. For more details see vignette:
\code{vignette("intro", package = "facefuns")}
}
