% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_arrays.R
\name{bind_arrays}
\alias{bind_arrays}
\title{Bind arrays}
\usage{
bind_arrays(...)
}
\arguments{
\item{...}{Matrices/arrays to be bound, separated by commas}
}
\value{
Returns new array
}
\description{
Binds two or more p x k matrices or p x k x n arrays into one three-dimensional array. Dimension names are set by those of first array/matrix entered
}
\note{
See \link[Morpho]{bindArr} for more flexible function that allows to bind along different dimensions
}
\examples{

data(londonSet_aligned)

new_array <- bind_arrays(LondonSet_aligned, LondonSet_aligned)

}
