% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAVG.R
\name{calcAVG}
\alias{calcAVG}
\title{Calculate averageness}
\usage{
calcAVG(quickstart_obj)
}
\arguments{
\item{quickstart_obj}{Quickstart object}
}
\value{
Returns tibble with distinctivness and averageness (reversed distinctiveness) scores
}
\description{
Functions calculates averageness/distinctiveness as each template's distance from sample average
}
