% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTEMtoTPS.R
\name{convertTEMtoTPS}
\alias{convertTEMtoTPS}
\title{Convert Webmorph template files to .tps format}
\usage{
convertTEMtoTPS(path_to_tem, p, remove_points = NA, path_to_tps)
}
\arguments{
\item{path_to_tem}{Path to directory containing template files}

\item{p}{Number of landmarks images were delineated with}

\item{remove_points}{Vector containing number of landmarks to be deleted. If no value is entered, all landmarks are retained}

\item{path_to_tps}{Name and path of TPS file to be created}
}
\description{
For more details see the corresponding vignette:
\code{vignette("TEMtoTPS", package = "facefuns")}
}
