% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2DPCs.R
\name{pc_pair_plot}
\alias{pc_pair_plot}
\title{pc_pair_plot}
\usage{
pc_pair_plot(f1, f2, pc)
}
\arguments{
\item{f1}{function for first plot}

\item{f2}{function for second plot}

\item{pc}{Numeric label}
}
\value{
plot
}
\description{
set up specs for plotting pairs of PCs
}
