% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pcs.R
\name{print.selected_PCs}
\alias{print.selected_PCs}
\title{Print for selected_PCs}
\usage{
\method{print}{selected_PCs}(x, ...)
}
\arguments{
\item{x}{a list of class selected_PCs}

\item{...}{arguments passed to or from other methods}
}
\value{
prints the selected PCs
}
\description{
Print for selected_PCs
}
\keyword{internal}
