% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lmdata.R
\name{read_lmdata}
\alias{read_lmdata}
\title{Read landmark data}
\usage{
read_lmdata(
  lmdata,
  specID = "None",
  remove_points = NA,
  path_to_tps = NA,
  plot = TRUE
)
}
\arguments{
\item{lmdata}{Landmark data as webmorph_list, tps file or tem file(s)}

\item{specID}{If reading TPS file, specify specID (see \code{geomorphs}'s \link[geomorph]{readland.tps})}

\item{remove_points}{If reading tem file(s), specify if any points should be removed}

\item{path_to_tps}{If reading temlist or tem file(s), specify if you want to save data as TPS file}

\item{plot}{Plot all specimens to check data (see \code{geomorphs}'s \link[geomorph]{plotAllSpecimens})}
}
\value{
Returns three-dimensional array of dimensions p, k, and n. p = number of landmarks, k = dimension (2D or 3D), n = number of specimens.
}
\description{
Read landmark data
}
\examples{
path_to_tem <- system.file("extdata", "tem", package="facefuns")
remove_points <- c(45:50, 100:104, 116:125, 146:158, 159:164, 165:170, 171:174, 175:179, 184:185)

data <- read_lmdata(lmdata =  path_to_tem, remove_points = remove_points)

}
